+++ plugins/Actions/Archiver.php
@@ -8,6 +8,7 @@
  */
 namespace Piwik\Plugins\Actions;
 
+use Piwik\Config;
 use Piwik\DataTable;
 use Piwik\Metrics as PiwikMetrics;
 use Piwik\RankingQuery;
@@ -425,6 +426,11 @@ class Archiver extends \Piwik\Plugin\Archiver
     protected function insertPageUrlsReports()
     {
         $dataTable = $this->getDataTable(Action::TYPE_PAGE_URL);
+
+        $prefix = $this->getProcessor()->getParams()->getSite()->getMainUrl();
+        $prefix = rtrim($prefix, '/') . '/';
+        $this->setFolderPathMetadata($dataTable, $isUrl = true, $prefix);
+
         $this->insertTable($dataTable, self::PAGE_URLS_RECORD_NAME);
 
         $records = array(
@@ -478,6 +484,7 @@ class Archiver extends \Piwik\Plugin\Archiver
     protected function insertPageTitlesReports()
     {
         $dataTable = $this->getDataTable(Action::TYPE_PAGE_TITLE);
+        $this->setFolderPathMetadata($dataTable, $isUrl = false);
         $this->insertTable($dataTable, self::PAGE_TITLES_RECORD_NAME);
     }
 
@@ -541,5 +548,24 @@ class Archiver extends \Piwik\Plugin\Archiver
 
         // Unique Keywords can't be summed, instead we take the RowsCount() of the keyword table
         $this->getProcessor()->insertNumericRecord(self::METRIC_KEYWORDS_RECORD_NAME, $nameToCount[self::SITE_SEARCH_RECORD_NAME]['level0']);
+    }
+
+    private function setFolderPathMetadata(DataTable $dataTable, $isUrl, $prefix = '')
+    {
+        $configGeneral = Config::getInstance()->General;
+        $separator = $isUrl ? '/' : $configGeneral['action_title_category_delimiter'];
+        $metadataName = $isUrl ? 'folder_url_start' : 'page_title_path';
+
+        foreach ($dataTable->getRows() as $row) {
+            $subtable = $row->getSubtable();
+            if (!$subtable) {
+                continue;
+            }
+
+            $metadataValue = $prefix . $row->getColumn('label');
+            $row->setMetadata($metadataName, $metadataValue);
+
+            $this->setFolderPathMetadata($subtable, $isUrl, $metadataValue . $separator);
+        }
     }
 }