+++ plugins/CoreConsole/Commands/GenerateAngularDirective.php
@@ -9,6 +9,7 @@
 
 namespace Piwik\Plugins\CoreConsole\Commands;
 
+use Piwik\Plugin\Manager;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -39,7 +40,7 @@ class GenerateAngularDirective extends GenerateAngularConstructBase
             throw new \Exception('The AngularJS directive ' . $directiveLower . ' already exists in plugin ' . $pluginName);
         }
 
-        $exampleFolder = PIWIK_INCLUDE_PATH . '/plugins/ExamplePlugin';
+        $exampleFolder = Manager::getPluginDirectory('ExamplePlugin');
         $replace       = array(
             'ExamplePlugin'       => $pluginName,
             'directive-component' => $directiveLower,