+++ plugins/Diagnostics/config/config.php
@@ -1,5 +1,7 @@
 <?php
 
+use Piwik\Plugins\Diagnostics\Diagnostic\CronArchivingLastRunCheck;
+
 return array(
     // Diagnostics for everything that is required for Piwik to run
     'diagnostics.required' => array(
@@ -23,6 +25,7 @@ return array(
         DI\get('Piwik\Plugins\Diagnostics\Diagnostic\RecommendedFunctionsCheck'),
         DI\get('Piwik\Plugins\Diagnostics\Diagnostic\NfsDiskCheck'),
         DI\get('Piwik\Plugins\Diagnostics\Diagnostic\CronArchivingCheck'),
+        DI\get(CronArchivingLastRunCheck::class),
         DI\get('Piwik\Plugins\Diagnostics\Diagnostic\LoadDataInfileCheck'),
         Di\get('Piwik\Plugins\Diagnostics\Diagnostic\DbOverSSLCheck'),
         Di\get('Piwik\Plugins\Diagnostics\Diagnostic\DbMaxPacket'),