+++ plugins/PrivacyManager/tests/Fixtures/MultipleSitesMultipleVisitsFixture.php
@@ -322,6 +322,7 @@ class MultipleSitesMultipleVisitsFixture extends Fixture
     {
         for ($day = 0; $day < $numIterations; $day++) {
             // we track over several days to make sure we have some data to aggregate in week reports
+            // NOTE: some action times are out of order in visits on purpose
 
             if ($day > 0) {
                 $this->trackingTime = Date::factory($this->dateTime)->addDay($day * 3)->getDatetime();